+++ HISTORY.md
@@ -1,3 +1,8 @@
+unreleased
+==========
+
+  * Remove object key-value-pair escape behavior
+
 2.3.3 / 2022-03-06
 ==================